Tests/LibWeb/Text/expected/hit_testing/content-editable-non-empty-block-descendants.txt

first:5
second:6
first:5
